requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
fastapi==0.112.2
filelock==3.15.4
h11==0.14.0
idna==3.8
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.3
python-dotenv==1.0.0
boto3
pandas
scikit-learn
joblib
